js/xpconnect/tests/mochitest/test_finalizationGroup.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test FinalizationGroup works in the browser</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript">
      function go() {
        SimpleTest.waitForExplicitFinish();

        // Group with no registered objects.
        let holdings1 = [];
        let group1 = new FinalizationGroup(i => holdings1 = [...i]);

        // Group with three registered objects.
        let holdings2 = [];
        let group2 = new FinalizationGroup(i => holdings2 = [...i]);
        group2.register({}, 1);
        group2.register({}, 2);
        group2.register({}, 3);

        // Group with registered object that is then unregistered.
        let holdings3 = [];
        let group3 = new FinalizationGroup(i => holdings3 = [...i]);
        let token3 = {}
        group3.register({}, 1, token3);
        group3.unregister(token3);

        // Group with registered object that doesn't die.
        let holdings4 = [];
        let group4 = new FinalizationGroup(i => holdings4 = [...i]);
        let object4 = {};
        group4.register(object4, 1);

        // Group observing cyclic JS data structure.
        let holdings5 = [];
        let group5 = new FinalizationGroup(i => holdings5 = [...i]);
        group5.register(makeJSCycle(4), 5);

        // Group observing cyclic DOM/JS data structure.
        let holdings6 = [];
        let group6 = new FinalizationGroup(i => holdings6 = [...i]);
        group6.register(makeDOMCycle(4), 6);

        // Need to run full GC/CC/GC cycle to collect cyclic garbage through DOM
        // and JS heaps.
        SpecialPowers.DOMWindowUtils.garbageCollect();
        SpecialPowers.DOMWindowUtils.cycleCollect();
        SpecialPowers.DOMWindowUtils.garbageCollect();

        // Microtasks are run before cleanup callbacks.
        Promise.resolve().then(() => {
          is(holdings1.length, 0);
          is(holdings2.length, 0);
          is(holdings3.length, 0);
          is(holdings4.length, 0);
          is(holdings5.length, 0);
          is(holdings6.length, 0);
        });

        // setTimeout queues a task which will run after cleanup callbacks.
        setTimeout(() => {
          is(holdings1.length, 0);

          let result = holdings2.sort((a, b) => a - b);
          is(result.length, 3);
          is(result[0], 1);
          is(result[1], 2);
          is(result[2], 3);

          is(holdings3.length, 0);
          is(holdings4.length, 0);

          is(holdings5.length, 1);
          is(holdings5[0], 5);

          is(holdings6.length, 1);
          is(holdings6[0], 6);

          SimpleTest.finish();
        }, 0);
      }

      function makeJSCycle(size) {
        let first = {};
        let current = first;
        for (let i = 0; i < size; i++) {
          current.next = {};
          current = current.next;
        }
        current.next = first;
        return first;
      }

      function makeDOMCycle(size) {
        let first = {};
        let current = first;
        for (let i = 0; i < size; i++) {
          if (i % 2 === 0) {
            current.next = document.createElement("div");
          } else {
            current.next = {};
          }
          current = current.next;
        }
        current.next = first;
        return first;
      }
    </script>
  </head>
  <body onload="go()"></body>
</html>
